+++ app/src/main/java/com/supersonic/heartrate/models/OnboardingPage.kt
@@ -3,5 +3,6 @@ package com.supersonic.heartrate.models
 data class OnboardingPage(
     val imageRes: Int,
     val titleRes: Int,
-    val bodyRes: Int
+    val bodyRes: Int,
+    val cameraPermission: Boolean = false
 )
