+++ app/src/main/res/values/strings.xml
@@ -5,7 +5,7 @@
     <string name="onboardingPage1_body">Оберіть точність з якою буде зроблене вимірювання пульсу. Чим вища точність тим точніший результат!</string>
 
     <string name="onboardingPage2_title">Початок вимірювання</string>
-    <string name="onboardingPage2_body">Щільно прикладіть палець до камери, так щоб повністю її закрити. Тримайте палець до закінчення процесу вимірювання пульсу.</string>
+    <string name="onboardingPage2_body">Надайте доступ до камери, за допомогою неї буде проводитись вимірювання вашого пульсу. Щільно прикладіть палець до камери, так щоб повністю її закрити. Тримайте палець до закінчення процесу вимірювання пульсу.</string>
 
     <string name="onboardingPage3_title">Результати</string>
     <string name="onboardingPage3_body">Після закінчення вимірювання ви отримаєте свій результат. Ви можете побачити минулі вимірювання на сторінці "Історія"</string>
